Now Services and Resources are seperated
