+++ src/main/java/org/lithium/Resources/TwitterPublishResource.java
@@ -1,7 +1,10 @@
-package org.lithium;
+package org.lithium.Resources;
 
 import com.codahale.metrics.annotation.Timed;
 import org.eclipse.jetty.util.log.Slf4jLog;
+import org.lithium.Response;
+import org.lithium.Services.TwitterService;
+import org.lithium.Tweet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import twitter4j.Status;
@@ -9,32 +12,34 @@ import twitter4j.Twitter;
 import twitter4j.TwitterException;
 import twitter4j.TwitterFactory;
 
-import javax.ws.rs.*;
+import javax.ws.rs.Consumes;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
-import java.util.List;
 import java.util.UUID;
 
-
-@Path("/timeline")
-@Produces(MediaType.APPLICATION_JSON)
-public class TwitterTimelineResource {
+@Path("/publishTweet")
+public class TwitterPublishResource {
     private Twitter twitter;
     private static Logger logger = LoggerFactory.getLogger(Slf4jLog.class);
-    public  TwitterTimelineResource(Twitter twitter) {
+    public  TwitterPublishResource(Twitter twitter) {
         this.twitter = twitter;
     }
-    @GET
+    @POST
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Produces(MediaType.APPLICATION_JSON)
     @Timed
-    public Response getTimeline() {
-        logger.info("Getting timeline");
+    public Response tweetMessage(Tweet tweet) {
+        logger.info("Publishing tweet: " + tweet);
         Response result = new Response();
         result.setGuid(UUID.randomUUID().toString());
         result.setSuccess(Boolean.FALSE);
         try {
-            List<Status> statuses = twitter.getHomeTimeline();
-            result.setTweets(statuses);
+            Status status = TwitterService.getInstance().postTweet(twitter, tweet.getMessage());
+            result.setMessage("Successfully updated the status to [" + status.getText() + "].");
             result.setSuccess(Boolean.TRUE);
-            logger.info("Got timeline.");
+            logger.info("Publishing successful");
         } catch (TwitterException e) {
             logger.error(e.getMessage(), e);
             result.setMessage(e.getMessage());
